+++ dev/preact/src/pages/Search/Content/Paginated.tsx
@@ -1,10 +1,39 @@
 import { SearchInput } from "@nosto/search-js/preact/autocomplete"
-import { useActions, useNostoAppState } from "@nosto/search-js/preact/hooks"
+import { speechToTextSupported, useActions, useNostoAppState, useSpeechToText } from "@nosto/search-js/preact/hooks"
 import { useState } from "preact/hooks"
 
 import { Input } from "../../../components/Input"
 import { Pagination } from "../../../components/Pagination"
 import { Product } from "../Product"
+
+function SpeechToTextButton() {
+  const { newSearch } = useActions()
+  const { startListening, listening, stopListening } = useSpeechToText()
+
+  if (!speechToTextSupported) {
+    return null
+  }
+
+  return (
+    <button
+      onClick={() => {
+        if (listening) {
+          stopListening()
+        } else {
+          startListening({
+            onResult: value => {
+              newSearch({
+                query: value
+              })
+            }
+          })
+        }
+      }}
+    >
+      🎤︎︎
+    </button>
+  )
+}
 
 export function Paginated() {
   const [input, setInput] = useState("")
@@ -28,7 +57,9 @@ export function Paginated() {
             componentProps={{ placeholder: "Search" }}
             onSearchInput={target => setInput(target.value)}
           />
+          <input type="search" placeholder="Search" />
           <input type="button" value="Search" onClick={onSearch} />
+          <SpeechToTextButton />
         </div>
       </div>
       <div>