+++ packages/core/src/resultCaching.ts
@@ -2,20 +2,22 @@ import { SearchQuery, SearchResult } from "@nosto/nosto-js/client"
 import { isEqual } from "@utils/isEqual"
 import { getSessionStorageItem, setSessionStorageItem } from "@utils/storage"
 
-export const STORAGE_ENTRY_NAME = "nosto:search:searchResult"
+export const STORAGE_ENTRY_NAME = "nosto:search-js:cache"
 
-export type SearchResultDto = {
+const TTL = 60 * 1000 // 1 minute
+
+type CacheEntry = {
   query: SearchQuery
   result: SearchResult
+  created: number
 }
 
 export function cacheSearchResult(query: SearchQuery, result: SearchResult) {
-  const dto: SearchResultDto = { query, result }
-  setSessionStorageItem(STORAGE_ENTRY_NAME, dto)
+  setSessionStorageItem(STORAGE_ENTRY_NAME, { query, result, created: Date.now() })
 }
 
 export function loadCachedResult(query: SearchQuery) {
-  const storageValue = getSessionStorageItem<SearchResultDto>(STORAGE_ENTRY_NAME)
+  const storageValue = getSessionStorageItem<CacheEntry>(STORAGE_ENTRY_NAME)
   if (!storageValue || !isValueShapeCorrect(storageValue)) {
     return null
   }
@@ -24,6 +26,10 @@ export function loadCachedResult(query: SearchQuery) {
   if (!isEqual(getCacheKey(query), cachedQuery)) {
     return null
   }
+  if (Date.now() - storageValue.created > TTL) {
+    return null
+  }
+
   return storageValue.result
 }
 
@@ -43,6 +49,6 @@ function getCacheKey(query: SearchQuery): SearchQuery {
 }
 
 // TODO: Better validation with valibot
-export function isValueShapeCorrect(value: unknown): value is SearchResultDto {
-  return typeof value === "object" && value !== null && "query" in value && "result" in value
+function isValueShapeCorrect(value: unknown): value is CacheEntry {
+  return typeof value === "object" && value !== null && "query" in value && "result" in value && "created" in value
 }